removed duplicated entities
