--- Code/Server/src/main/java/model/CustomerInfoEntity.java
@@ -1,90 +1,0 @@
-package model;
-
-import javax.persistence.*;
-
-/**
- * Created by homepppp on 2017/6/27.
- */
-@Entity
-@Table(name = "customerinfo", schema = "tradein", catalog = "")
-public class CustomerInfoEntity {
-    private int userId;
-    private String email;
-    private String phone;
-    private Integer rating;
-    private UserEntity usersByUserId;
-
-    @Id
-    @Column(name = "userID", nullable = false)
-    public int getUserId() {
-        return userId;
-    }
-
-    public void setUserId(int userId) {
-        this.userId = userId;
-    }
-
-    @Basic
-    @Column(name = "email", nullable = true, length = 255)
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    @Basic
-    @Column(name = "phone", nullable = true, length = 127)
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    @Basic
-    @Column(name = "rating", nullable = true)
-    public Integer getRating() {
-        return rating;
-    }
-
-    public void setRating(Integer rating) {
-        this.rating = rating;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        CustomerInfoEntity that = (CustomerInfoEntity) o;
-
-        if (userId != that.userId) return false;
-        if (email != null ? !email.equals(that.email) : that.email != null) return false;
-        if (phone != null ? !phone.equals(that.phone) : that.phone != null) return false;
-        if (rating != null ? !rating.equals(that.rating) : that.rating != null) return false;
-
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        int result = userId;
-        result = 31 * result + (email != null ? email.hashCode() : 0);
-        result = 31 * result + (phone != null ? phone.hashCode() : 0);
-        result = 31 * result + (rating != null ? rating.hashCode() : 0);
-        return result;
-    }
-
-    @OneToOne
-    @JoinColumn(name = "userID", referencedColumnName = "userID", nullable = false)
-    public UserEntity getUsersByUserId() {
-        return usersByUserId;
-    }
-
-    public void setUsersByUserId(UserEntity usersByUserId) {
-        this.usersByUserId = usersByUserId;
-    }
-}
