+++ Code/Server/src/main/java/model/OfferEntityPK.java
@@ -1,8 +1,6 @@
 package model;
 
-import javax.persistence.Column;
 import javax.persistence.Embeddable;
-import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 import java.io.Serializable;
 
@@ -11,15 +9,15 @@ import java.io.Serializable;
  */
 @Embeddable
 public class OfferEntityPK implements Serializable {
-    private GamesEntity game;
+    private GameEntity game;
     private UserEntity user;
 
     @ManyToOne
-    public GamesEntity getGame() {
+    public GameEntity getGame() {
         return game;
     }
 
-    public void setGame(GamesEntity game) {
+    public void setGame(GameEntity game) {
         this.game = game;
     }
